refactor: Remove product categories and change field names

+++ be/api/v1/models/cart.py
@@ -6,13 +6,13 @@ class CartItem(BaseModel):
     size: Optional[str]
     quantity: int
     colorway: str
-    
+
 class Cart(BaseModel):
     items: list[CartItem]
     promoCode: Optional[str]
 
 class PriceModel(BaseModel):
     currency: str
-    subtotal: float
-    discount: float
-    grandTotal: float
+    subtotal: int
+    discount: int
+    grandTotal: int
